fix: one more attempt

+++ core/network/impl/peer_manager_impl.cpp
@@ -112,7 +112,7 @@ namespace kagome::network {
     }
   }
 
-  size_t PeerManagerImpl::activePeersNumber() const {
+  uint32_t PeerManagerImpl::activePeersNumber() const {
     return active_peers_.size();
   }
